fixing make bugs

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: csiqueir <carlos.felipe@hotmail.com.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/13 20:19:43 by csiqueir          #+#    #+#             */
-/*   Updated: 2022/05/14 11:49:03 by csiqueir         ###   ########.fr       */
+/*   Updated: 2022/05/25 19:23:20 by csiqueir         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	int index;
+	size_t index;
 
 index = 0;
 if (dstsize != 0 )
